feat: 'runMPS' tasks to execute code in an MPS instance
In the 'mpsBuild' section, you can now use 'runMPS' to specify a task
that launches a headless MPS instance and executes a given method.

MPS provides an ANT task for running a static method, but it has to be
part of a solution. This plugin generates such a solution with all the
dependencies and jar libraries configured.